requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
chromium==0.0.0
click==8.1.8
cloudinary==1.44.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
edge-tts==7.0.2
fastapi==0.116.1
faster-whisper==1.1.1
ffmpeg-python==0.2.0
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.74.0
grpcio-status==1.74.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
humanfriendly==10.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-google-genai==2.1.8
langchain-text-splitters==0.3.9
langgraph==0.5.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.74
langsmith==0.4.8
marshmallow==3.26.1
motor==3.7.1
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
numpy==2.3.2
onnxruntime==1.22.1
openai==1.97.1
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
playwright==1.54.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
pymongo==4.13.2
pyreadline3==3.5.4
python-dotenv==1.1.1
PyYAML==6.0.2
RapidFuzz==3.13.0
redis==6.2.0
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
SQLAlchemy==2.0.41
srt==3.5.3
starlette==0.47.2
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tinydb==4.8.2
tokenizers==0.21.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
